fix hmin/hmax of exprs
